+++ lib/presentation/router/app_router.dart
@@ -6,13 +6,14 @@ import 'package:flutter_practice/presentation/screens/calculator_screen.dart';
 import 'package:flutter_practice/presentation/screens/home_screen.dart';
 import 'package:flutter_practice/presentation/screens/splash_screen.dart';
 import 'package:flutter_practice/presentation/screens/theme_setting_screen.dart';
+import 'package:flutter_practice/presentation/screens/voice_assistant_screen.dart';
 
 import '../../core/exceptions/route_exception.dart';
 
 class AppRouter {
   static const String splash = '/',
       home = 'home', calc = 'calc', randomJokes = 'randomJokes', myForm = 'myForm',
-      lang = 'lang', theme = 'theme';
+      lang = 'lang', theme = 'theme', va = 'va';
 
   const AppRouter._();
 
@@ -45,6 +46,10 @@ class AppRouter {
       case theme:
         return MaterialPageRoute(
           builder: (_) => const ThemeSettingScreen(),
+        );
+      case va:
+        return MaterialPageRoute(
+          builder: (_) => const VoiceAssistantScreen(),
         );
       default:
         throw const RouteException('Route not found!');